+++ Magna/src/main/java/de/morigm/magna/api/helper/StringHelper.java
@@ -12,5 +12,14 @@ public class StringHelper
 		return tmp;
 	}
 	
+	public static String StringArrayToString(String[] arr,String split,int count)
+	{
+		String tmp = "";
+		for(int i = count;i < arr.length;i++)
+			tmp += arr[i] + split;
+		tmp = tmp.substring(0, (tmp.length() - split.length()));
+		return tmp;
+	}
+	
 	
 }