+++ Magna/src/main/java/de/morigm/magna/stuff/MagnaStuff.java
@@ -1,0 +1,17 @@
+package de.morigm.magna.stuff;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import lombok.Getter;
+
+public class MagnaStuff 
+{
+	static
+	{
+		msg = new HashMap<>();
+	}
+	
+	
+	@Getter private static Map<String,String> msg;
+}